requirements.txt
python-telegram-bot>=13.0
google-generativeai>=0.1.0
python-dotenv>=0.19.0
requests>=2.26.0
Flask>=2.0.0
gunicorn>=20.1.0